techchallenge/requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Django==5.1.1
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.8
inflection==0.5.1
jwcrypto==1.5.6
mypy-extensions==1.0.0
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.3
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.26
